backend/datagib/requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
dateparser==0.7.0
Deprecated==1.2.4
Django==2.1.4
django-filter==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.9.0
google-api-core==1.6.0
google-auth==1.6.1
google-cloud-bigquery==1.8.1
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
google-cloud-language==1.1.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
grpcio==1.17.1
humanize==0.5.1
idna==2.8
Markdown==3.0.1
maya==0.5.0
pendulum==1.5.1
protobuf==3.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyGithub==1.43.4
PyJWT==1.7.1
python-dateutil==2.7.5
pytz==2018.7
pytzdata==2018.7
regex==2018.11.22
requests==2.21.0
rsa==4.0
sentry-sdk==0.6.2
six==1.12.0
snaptime==0.2.4
tzlocal==1.5.1
urllib3==1.24.1
wrapt==1.11.1
